Medic_AI/files/medical_chatbot_project/requirements.txt
# Medical AI Chatbot Dependencies

# Core LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0

# Vector Database
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0

# Web Interface
gradio>=4.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tiktoken>=0.5.2

# Optional but recommended
openai>=1.10.0
anthropic>=0.8.0
